+++ aoe4/aoe4_settings.py
@@ -118,10 +118,10 @@ class AoE4Hotkeys(SettingsSubclass):
         """Constructor"""
         self.enter: str = 'Return'  # enter selection key
         self.select_next_build_order: str = 'Tab'  # select the next build order
-        self.next_panel: str = '<ctrl>+<alt>+q'  # cycle through the next panel
-        self.show_hide: str = '<ctrl>+<alt>+w'  # show/hide the application
-        self.build_order_previous_step: str = '<ctrl>+<alt>+e'  # go to the previous build order step
-        self.build_order_next_step: str = '<ctrl>+<alt>+r'  # go to the next build order step
+        self.next_panel: str = ''  # cycle through the next panel
+        self.show_hide: str = ''  # show/hide the application
+        self.build_order_previous_step: str = ''  # go to the previous build order step
+        self.build_order_next_step: str = ''  # go to the next build order step
 
 
 class AoE4OverlaySettings(SettingsSubclass):