+++ Changelog.md
@@ -1,7 +1,10 @@
+## [1.4.0] - in progress
+* Hotkeys can be updated by reloading the settings.
+
 ## [1.3.4] - 2022.09.11
 * Resource targets can be split between different resource subtypes, by hovering the mouse.
     * A tooltip appears on a separate window (if build order adapted for it).
-* The seach bar for the build order performs a fuzzy search to find the best build orders. 
+* The search bar for the build order performs a fuzzy search to find the best build orders. 
 * AoE2
     * Villagers and age only indicated if information available.
 * AoE4